pyproject.toml
[project]
name = "vscode"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "aiohttp>=3.13.2",
    "asyncpg>=0.31.0",
    "cryptography>=46.0.6",
    "discord-py>=2.6.4",
    "psutil>=7.1.3",
    "pydantic>=2.12.5",
    "pydantic-settings>=2.13.1",
    "python-dotenv>=1.2.1",
    "redis>=7.1.0",
    "websockets>=15.0.1",
]
